Pip macos name (#1)
* Pip: fix operating system name for macOS

The code currently inputs operatingSystem of macOS as "mac" when running python-inspector.

However, python-inspector expects the option of macOS to be "macos".

Signed-off-by: Michael Yoo <vbtkdpf148@gmail.com>

* JitPack: Switch back to OpenJDK 11

Switch back to OpenJDK 11 for building the JitPack artifacts. Otherwise
it is not possible to use those artifacts to build plugins that still
need to run on Java 11, for example when using a customer Docker image
with Java 11 instead of 17.

Trying to build with Java 17 and target 11 results in this error
message:

Cannot inline bytecode built with JVM target 17 into bytecode that is
being built with JVM target 11. Please specify proper '-jvm-target'
option.

This partly reverts 171fa3d.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>

* AnalyzerResultBuilder: Remove a duplicate word

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>

* CreateAnalyzerResultCommand: Extract a variable

Reduce the length of the long call chain a bit.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Use `orEmpty()` to eliminate two `?`

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Only use package based storages

The code was written back when provenance based storages hadn't existed
yet. This check was just forgotten to add when provenance based storages
had been added.

Note: The implementation only works with a package based storage.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Migrate to provenance based scan storage

The command operates on a package based scan result storage which has
become legacy. So, operate on a provenance based scan result storage
instead which has become the default way of storing scan results.

The most recent resolved provenance now is no more determined via the
scan start time, but via the row id of the provenance resultion result.
This is much faster, because it works without accessing the scan result
JSONB column.

Note: The primary use case this command targets is re-scanning packages
prior to upgrading ScanCode so that the "cache" is warmed-up.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Slightly improve a query

Select only distinct entries in order to not process any duplicates.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Limit the query to `ScanCode` entries

Align with the documentation of the option `--scancode-version`,

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* CreateAnalyzerResultCommand: Slightly improve the help text

Signed-off-by: Frank Viernau <frank_viernau@epam.com>

* Pip: change configuration key for macOS

Configuration key of macOS for python-inspector is "macos", whereas in Pip.kt it is set to "mac".
So, change the configuration key to "macos".

Signed-off-by: Michael Yoo <vbtkdpf148@gmail.com>

Signed-off-by: Michael Yoo <vbtkdpf148@gmail.com>
Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>
Signed-off-by: Frank Viernau <frank_viernau@epam.com>
Co-authored-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>
Co-authored-by: Frank Viernau <frank_viernau@epam.com>